Fixed Table component and CRUD operations for the various pages that use the table (#28)
this took FAR FAR too long to finish but it's finally here!!

---------

+++ backend/api/routes/resource_slots.py
@@ -9,6 +9,7 @@ from sqlalchemy.orm import InstrumentedAttribute, selectinload
 from models.audit_log import AuditLog
 from models.resource import Resource
 from models.resource_slot import ResourceSlot
+from models.machine_type import MachineType, MachineTypeSlotAssociation
 from models.resource_usage_quantity import ResourceUsageQuantity
 from schemas.requests import ResourceSlotCreateRequest, ResourceSlotEditRequest
 from schemas.responses import (
@@ -117,6 +118,7 @@ async def get_resource_slot(
     return ResourceSlotDetails(
         audit_logs=[AuditLogModel.model_validate(log) for log in audit_logs],
         valid_resource_ids=[resource.id for resource in resource_slot.valid_resources],
+        resource_names={resource.name for resource in resource_slot.valid_resources},
         **resource_slot.__dict__,
     )
 
@@ -164,6 +166,9 @@ async def get_all_resource_slots(
             valid_resource_ids=[
                 resource.id for resource in resource_slot.valid_resources
             ],
+            resource_names={
+                resource.name for resource in resource_slot.valid_resources
+            },
             **resource_slot.__dict__,
         )
         for resource_slot in resource_slots
@@ -207,8 +212,8 @@ async def edit_resource_slot(
             if resource_slot.display_name != request.display_name
             else None
         ),
-        "valid_resource_ids": (
-            request.resource_ids
+        "resource_ids": (
+            [str(r) for r in request.resource_ids]
             if set(resource_slot.valid_resources) != set(valid_resources)
             else None
         ),
@@ -260,6 +265,14 @@ async def delete_resource_slot(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Resource Slot with provided ID not found",
+        )
+
+    slotAcc: MachineTypeSlotAssociation = await session.scalar(select(MachineTypeSlotAssociation).where(MachineTypeSlotAssociation.resource_slot_id == resource_slot_id))
+    if slotAcc:
+        mtypeWith: MachineType = await session.get(MachineType, slotAcc.machine_type_id)
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"One or more machine types is using this resource slot: {mtypeWith.name})"
         )
 
     await session.delete(resource_slot)
